+++ libraries/mapillary/src/components/Mapillary/designer.ts
